+++ lib/iris/tests/unit/config/test_Parallel.py
@@ -14,7 +14,7 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with Iris.  If not, see <http://www.gnu.org/licenses/>.
-"""Unit tests for the `iris.options.Paralle` class."""
+"""Unit tests for the :class:`iris.config.Parallel` class."""
 
 from __future__ import (absolute_import, division, print_function)
 from six.moves import (filter, input, map, range, zip)  # noqa
@@ -30,7 +30,7 @@ import warnings
 import dask
 import distributed
 
-from iris.options import Parallel
+from iris.config import Parallel
 from iris.tests import mock
